refactor(protocol): split the database port into entities, query groups, and capability views
